+++ include/ccf/historical_queries_interface.h
@@ -7,62 +7,13 @@
 #include "consensus/ledger_enclave_types.h"
 #include "kv/store.h"
 #include "node/history.h"
-#include "tls/base64.h"
+#include "node/tx_receipt.h"
 
 #include <chrono>
 #include <memory>
 
 namespace ccf::historical
 {
-  // TODO: Move out of historical namespace
-  struct TxReceipt
-  {
-    std::vector<uint8_t> signature = {};
-    HistoryTree::Hash root = {};
-    std::shared_ptr<ccf::HistoryTree::Path> path = {};
-    ccf::NodeId node_id = {};
-
-    TxReceipt(
-      const std::vector<uint8_t>& s_,
-      const HistoryTree::Hash& r_,
-      std::shared_ptr<ccf::HistoryTree::Path> p_,
-      const NodeId& n_) :
-      signature(s_),
-      root(r_),
-      path(p_),
-      node_id(n_)
-    {}
-
-    void describe(ccf::Receipt& r)
-    {
-      r.signature = tls::b64_from_raw(signature);
-      r.root = root.to_string();
-      if (path)
-      {
-        for (const auto& node : *path)
-        {
-          ccf::Receipt::Element n;
-          if (node.direction == ccf::HistoryTree::Path::Direction::PATH_LEFT)
-          {
-            n.left = node.hash.to_string();
-          }
-          else
-          {
-            n.right = node.hash.to_string();
-          }
-          r.proof.emplace_back(std::move(n));
-        }
-        r.leaf = path->leaf().to_string();
-      }
-      else
-      {
-        r.leaf = r.root;
-      }
-      r.node_id = node_id;
-    }
-  };
-
-  using TxReceiptPtr = std::shared_ptr<TxReceipt>;
   using StorePtr = std::shared_ptr<kv::Store>;
 
   struct State
